--- designSystem/src/main/java/com/example/designsystem/colorStyle/BrandColors.kt
@@ -1,10 +1,0 @@
-package com.example.designsystem.colorStyle
-
-import androidx.compose.ui.graphics.Color
-
-data class BrandColors(
-    val primary: Color,
-    val onPrimary: Color,
-    val primaryContainer: Color,
-    val onPrimaryContainer: Color,
-)
