--- designSystem/src/main/java/com/example/designsystem/colorStyle/SurfaceColors.kt
@@ -1,19 +1,0 @@
-package com.example.designsystem.colorStyle
-
-import androidx.compose.ui.graphics.Color
-
-data class SurfaceColor(
-    val surface: Color,
-    val onSurface: Color,
-    val surfaceContainer: Color,
-    val onSurfaceContainer: Color,
-    val surfaceVariant: Color,
-    val onSurfaceVariant: Color,
-    val outline: Color,
-    val outlineVariant: Color,
-    val onSurface_1: Color,
-    val onSurface_2: Color,
-    val onSurface_3: Color,
-    val onSurface_4: Color,
-)
-
