--- designSystem/src/main/java/com/example/designsystem/colorStyle/SystemColors.kt
@@ -1,17 +1,0 @@
-package com.example.designsystem.colorStyle
-
-import androidx.compose.ui.graphics.Color
-
-data class SystemColors(
-    val error: Color,
-    val onError: Color,
-    val errorContainer: Color,
-    val onErrorContainer: Color,
-    val warning: Color,
-    val onWarning: Color,
-    val onWarningContainer: Color,
-    val success: Color,
-    val onSuccess: Color,
-    val successContainer: Color,
-    val onSuccessContainer: Color
-)
