--- designSystem/src/main/java/com/example/designsystem/textStyle/BodyTextStyle.kt
@@ -1,10 +1,0 @@
-package com.example.designsystem.textStyle
-
-import androidx.compose.ui.text.TextStyle
-
-data class BodyTextStyle(
-    val smallRegular16: TextStyle,
-    val medium14: TextStyle,
-    val medium12: TextStyle,
-    val smallRegular10: TextStyle,
-)
