--- designSystem/src/main/java/com/example/designsystem/textStyle/DisplayTextStyle.kt
@@ -1,10 +1,0 @@
-package com.example.designsystem.textStyle
-
-import androidx.compose.ui.text.TextStyle
-
-data class DisplayTextStyle(
-    val largeBold24: TextStyle,
-    val largeBold20: TextStyle,
-    val mediumBold24: TextStyle,
-    val largeBold18: TextStyle,
-    )
