--- designSystem/src/main/java/com/example/designsystem/textStyle/HeadlineTextStyle.kt
@@ -1,10 +1,0 @@
-package com.example.designsystem.textStyle
-
-import androidx.compose.ui.text.TextStyle
-
-data class HeadlineTextStyle(
-    val largeBold18: TextStyle,
-    val medium18: TextStyle,
-    val largeBold16: TextStyle,
-    val medium16: TextStyle,
-)
