--- designSystem/src/main/java/com/example/designsystem/textStyle/LabelTextStyle.kt
@@ -1,11 +1,0 @@
-package com.example.designsystem.textStyle
-
-import androidx.compose.ui.text.TextStyle
-
-data class LabelTextStyle(
-    val largeMedium16: TextStyle,
-    val medium14: TextStyle,
-    val smallRegular14: TextStyle,
-    val medium12: TextStyle,
-    val smallRegular12: TextStyle,
-)
