--- designSystem/src/main/java/com/example/designsystem/textStyle/TitleTextStyle.kt
@@ -1,10 +1,0 @@
-package com.example.designsystem.textStyle
-
-import androidx.compose.ui.text.TextStyle
-
-data class TitleTextStyle(
-    val largeBold16: TextStyle,
-    val medium16: TextStyle,
-    val largeBold14: TextStyle,
-    val medium14: TextStyle,
-)
